// app/types/app.ts
export type ExploreDataType = {
    img: string,
    location: string,
    distance: string
}

export type LiveDataType = {
    img: string,
    title: string,
}

export type ListingCardItem = {
    img: string;
    location: string;
    title: string;
    description: string;
    star: number;
    price: string;
    total: string;
    long: number;
    lat: number;
};